backend/app/requirements.txt
-i https://pypi.org/simple
alembic==1.8.1
amqp==5.1.1 ; python_version >= '3.6'
anyio==3.6.2 ; python_full_version >= '3.6.2'
async-timeout==4.0.2 ; python_version >= '3.6'
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.3
deprecated==1.2.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
dnspython==2.2.1 ; python_version >= '3.6' and python_version < '4.0'
ecdsa==0.18.0 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==1.3.0
fastapi==0.86.0
greenlet==2.0.0.post0 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
httptools==0.5.0
idna==3.4 ; python_version >= '3.5'
kombu==5.2.4 ; python_version >= '3.7'
mako==1.2.3 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
packaging==21.3 ; python_version >= '3.6'
passlib[bcrypt]==1.7.4
pillow==9.3.0
prompt-toolkit==3.0.32 ; python_full_version >= '3.6.2'
psycopg2==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2 ; python_version >= '3.7'
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dotenv==0.21.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
pytz==2022.6
pyyaml==6.0
redis==4.3.4
rsa==4.9 ; python_version >= '3.6' and python_version < '4'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==1.4.41 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
sqlalchemy2-stubs==0.0.2a29 ; python_version >= '3.6'
sqlmodel==0.0.8
starlette==0.20.4 ; python_version >= '3.7'
typing-extensions==4.4.0 ; python_version >= '3.7'
uvicorn[standard]==0.19.0
uvloop==0.17.0
vine==5.0.0 ; python_version >= '3.6'
watchfiles==0.18.0
wcwidth==0.2.5
websockets==10.4
wrapt==1.14.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
